Remove blocking ReadFrom in threadSafeBuffer
Because ReadFrom will block on the src io.Reader it is passed, and
thereby hold the lock indefinitely, we don't want to have it as part
of the threadSafeBuffer.

Instead, don't embed the bytes.Buffer so as not to inherit ReadFrom,
and implement just the minimum needed for io.Reader and io.Writer.